+++ CHANGELOG.md
@@ -6,12 +6,12 @@
 
 ### 0.1.4
 
-* Fix generator issue with backwards compat (with ember-cli < 0.0.43). Previously, running `ember g divshot`
+* Fix generator issue with backwards compat (with ember-cli < 0.0.43). Previously, running `ember g parse`
   would error in 0.0.42.
 
 ### 0.1.3
 
-* Add `.divshot-cache` to `.gitignore` when running `ember g divshot`.
+* Add `.parse-cache` to `.gitignore` when running `ember g divshot`.
 
 ### 0.1.2